// client/src/app/api/model/raceFormatDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AthleteGroupDTO } from './athleteGroupDTO';


/**
 * Формат соревнования
 */
export interface RaceFormatDTO { 
    /**
     * Идентификатор формата
     */
    id: number;
    /**
     * Название формата
     */
    name: string;
    /**
     * Описание
     */
    description?: string;
    /**
     * Тип участия
     */
    type: string;
    /**
     * Состояние
     */
    state: string;
    /**
     * Дата-время начала
     */
    startTime?: string;
    /**
     * Дата-время завершения
     */
    finishTime?: string;
    /**
     * Группы участников
     */
    athleteGroups?: Array<AthleteGroupDTO>;
}